Create a package for terminal I/O operations
