+++ commands/add/phrase/phrase.go
@@ -11,6 +11,7 @@ import (
 	cmdutil "github.com/GGP1/kure/commands"
 	"github.com/GGP1/kure/db/entry"
 	"github.com/GGP1/kure/pb"
+	"github.com/GGP1/kure/terminal"
 
 	"github.com/GGP1/atoll"
 
@@ -90,10 +91,10 @@ func runPhrase(db *bolt.DB, r io.Reader, opts *phraseOptions) cmdutil.RunEFunc {
 func entryInput(r io.Reader, name string) (*pb.Entry, error) {
 	reader := bufio.NewReader(r)
 
-	username := cmdutil.Scanln(reader, "Username")
-	url := cmdutil.Scanln(reader, "URL")
-	expires := cmdutil.Scanln(reader, "Expires [dd/mm/yy]")
-	notes := cmdutil.Scanlns(reader, "Notes")
+	username := terminal.Scanln(reader, "Username")
+	url := terminal.Scanln(reader, "URL")
+	expires := terminal.Scanln(reader, "Expires [dd/mm/yy]")
+	notes := terminal.Scanlns(reader, "Notes")
 
 	exp, err := cmdutil.FmtExpires(expires)
 	if err != nil {
